+++ poetry/console/styles/poetry.py
@@ -9,7 +9,7 @@ class PoetryStyle(CleoStyle):
 
         self.output.get_formatter().add_style('error', 'red')
         self.output.get_formatter().add_style('warning', 'yellow')
-        self.output.get_formatter().add_style('question', 'blue')
+        self.output.get_formatter().add_style('question', 'cyan')
         self.output.get_formatter().add_style('comment', 'blue')
 
     def writeln(self, messages,