+++ poetry/layouts/layout.py
@@ -1,7 +1,6 @@
 from poetry.toml import dumps
 from poetry.toml import loads
 from poetry.utils.helpers import module_name
-from poetry.vcs.git import Git
 
 
 TESTS_DEFAULT = u"""from {package_name} import __version__
@@ -20,45 +19,52 @@ description = ""
 authors = []
 
 [tool.poetry.dependencies]
-python = "*"
 
 [tool.poetry.dev-dependencies]
-pytest = "^3.5"
+"""
+
+POETRY_WITH_LICENSE = """\
+[tool.poetry]
+name = ""
+version = ""
+description = ""
+authors = []
+license = ""
+
+[tool.poetry.dependencies]
+
+[tool.poetry.dev-dependencies]
 """
 
 
 class Layout(object):
 
-    def __init__(self, project, version='0.1.0', readme_format='md', author=None):
+    def __init__(self,
+                 project,
+                 version='0.1.0',
+                 description='',
+                 readme_format='md',
+                 author=None,
+                 license=None,
+                 python='*',
+                 dependencies=None,
+                 dev_dependencies=None):
         self._project = project
         self._package_name = module_name(project)
         self._version = version
+        self._description = description
         self._readme_format = readme_format
-        self._dependencies = {}
-        self._dev_dependencies = {}
-        self._include = []
+        self._license = license
+        self._python = python
+        self._dependencies = dependencies or {}
+        self._dev_dependencies = dev_dependencies or {'pytest': '^3.5'}
 
-        self._git = Git()
-        git_config = self._git.config
         if not author:
-            if (
-                git_config.get('user.name')
-                and git_config.get('user.email')
-            ):
-                author = u'{} <{}>'.format(
-                    git_config['user.name'],
-                    git_config['user.email']
-                )
-            else:
-                author = 'Your Name <you@example.com>'
+            author = 'Your Name <you@example.com>'
 
         self._author = author
 
     def create(self, path, with_tests=True):
-        self._dependencies = {}
-        self._dev_dependencies = {}
-        self._include = []
-
         path.mkdir(parents=True, exist_ok=True)
 
         self._create_default(path)
@@ -68,6 +74,30 @@ class Layout(object):
             self._create_tests(path)
 
         self._write_poetry(path)
+
+    def generate_poetry_content(self):
+        template = POETRY_DEFAULT
+        if self._license:
+            template = POETRY_WITH_LICENSE
+
+        content = loads(template)
+        poetry_content = content['tool']['poetry']
+        poetry_content['name'] = self._project
+        poetry_content['version'] = self._version
+        poetry_content['description'] = self._description
+        poetry_content['authors'].append(self._author)
+        if self._license:
+            poetry_content['license'] = self._license
+
+        poetry_content['dependencies']['python'] = self._python
+
+        for dep_name, dep_constraint in self._dependencies.items():
+            poetry_content['dependencies'][dep_name] = dep_constraint
+
+        for dep_name, dep_constraint in self._dev_dependencies.items():
+            poetry_content['dev-dependencies'][dep_name] = dep_constraint
+
+        return dumps(content)
 
     def _create_default(self, path, src=True):
         raise NotImplementedError()
@@ -99,13 +129,9 @@ class Layout(object):
             )
 
     def _write_poetry(self, path):
-        content = loads(POETRY_DEFAULT)
-        poetry_content = content['tool']['poetry']
-        poetry_content['name'] = self._project
-        poetry_content['version'] = self._version
-        poetry_content['authors'].append(self._author)
+        content = self.generate_poetry_content()
 
         poetry = path / 'pyproject.toml'
 
         with poetry.open('w') as f:
-            f.write(dumps(content))
+            f.write(content)